Cambio firma de estadoDelMar por que tiraba overflow en parametro

+++ Codigo/Evento_Pesca.bas
@@ -79,7 +79,7 @@ Public Function DamePez(ByVal ZonaUser As Byte) As Long
     DamePez = Zona(ZonaUser).Peces(RandomNumber(LBound(Zona(ZonaUser).Peces()), UBound(Zona(ZonaUser).Peces())))
 End Function
 
-Public Sub estadoDelMar(ByVal MinsEventoPesca As Byte)
+Public Sub estadoDelMar(ByVal MinsEventoPesca As Integer)
 
     With PescaEvent
